fix: clear the remaining generated-surface defects
Takes the sweep from 20 errors to 0, across five more causes.

Non-empty String, StringName and NodePath defaults were discarded. Only
the empty cases were special-cased; anything else fell through to
`parse`, yielding a nullable value that materialized as `.init()`.
`NodePath("")` additionally became `NodePath.copy("")`, and `copy` takes
a NodePath.

For String this was silent rather than a compile error:
`FileAccess.get_csv_line(delim: String = ",")` built fine and defaulted
to `""`, as did `bind_address = "*"` on four networking methods. No
compiler could have caught it; it turned up only from reading the API
next to the generated output while chasing the StringName failure.
Defaults that cannot be a comptime field initializer now carry an
explicit `Parameter.default_materializer`, which `Variant = 0` uses too.

Six `Variant` methods -- call, set, setNamed, setKeyed, setIndexed, ptr
-- took `self: *Variant` and then passed `&self`, handing the engine the
address of the parameter slot instead of the value. Their neighbours
take `self` by value, where `&self` is correct, which is likely how it
survived. Nothing referenced any of the six, so nothing compiled them.

Also: `variantGetObjectInstanceId` was called without unwrapping its
optional dispatch pointer; `Variant.Tag.forType(Variant)` hit a
compileError where the engine (and godot-cpp) use NIL for a dynamically
typed operand; `Rect2.initXYWidthHeight` took i64 and forwarded to a f64
constructor; and `Plane`'s constants called the engine-backed `initABCD`
from container scope, where a comptime-known initializer is required.

`vtable.zig`'s eval branch quota was not a defect: `combineNames`
computes its return type by calling `countNew`, evaluated before the
body and any quota it sets. Raised at `extend`, the outermost entry.

CI now runs the sweep as its own step so a surface regression reads
differently from a test failure. It stays opt-in locally: attempts to
measure the compile-time cost only ever produced cache hits, so the
conservative split was chosen rather than a guess.

+++ src/builtin/Rect2.mixin.zig
@@ -7,7 +7,7 @@ pub const init: Rect2 = .initPositionSize(
 /// Constructs a [Rect2](https://gdzig.github.io/gdzig/#gdzig.builtin.rect2.Rect2) by setting its `position` to (`x`, `y`), and its `size` to (`width`, `height`).
 ///
 /// @comptime
-pub fn initXYWidthHeight(p_x: i64, p_y: i64, p_width: i64, p_height: i64) Rect2 {
+pub fn initXYWidthHeight(p_x: f64, p_y: f64, p_width: f64, p_height: f64) Rect2 {
     return .initPositionSize(
         .initXY(p_x, p_y),
         .initXY(p_width, p_height),
